backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
supabase==2.0.2
openai==1.3.5
langchain==0.0.340
langchain-openai==0.0.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
